requirements.txt
amqp==2.2.2
astroid==1.5.3
autopep8==1.3.2
billiard==3.5.0.3
celery==4.1.0
Django==2.0.1
django-audiofield==0.10.0
django-celery==3.2.2
django-filter==1.1.0
djangorestframework==3.7.7
isort==4.2.15
kombu==4.1.0
lazy-object-proxy==1.3.1
Markdown==2.6.9
mccabe==0.6.1
psycopg2==2.7.3.1
pycodestyle==2.3.1
pylint==1.7.4
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
pytz==2017.3
six==1.11.0
vine==1.1.4
wrapt==1.10.11
